+++ exec/me_findjov.c
@@ -15,9 +15,9 @@
 void me_findjov(
                  long int mjd, /* (input) modified julian date */
                  long int mpm, /* (input) milliseconds past UTC midnight */ 
-                 float *ra,    /* (output) [h] RA */
-                 float *dec,   /* (output) [deg] dec */
-                 float *dist   /* (output) [AU] distance from Earth */
+                 double *ra,   /* (output) [h] RA */
+                 double *dec,  /* (output) [deg] dec */
+                 double *dist   /* (output) [AU] distance from Earth */
                 ) {
 
   double tai1, tai2, tt1, tt2;
@@ -64,10 +64,9 @@ void me_findjov(
   iauAb(tpv[0], astrom.v, astrom.em, astrom.bm1, &pv[0][0]);
   iauPv2s(pv, &L, &B, &R2, &dL, &dB, &dR);
   
-  /* Back to floats */
-  *ra = (float) L * DR2D / 15;
-  *dec = (float) B * DR2D;
-  *dist = (float) R;
+  *ra = L * (DR2D / 15);
+  *dec = B * DR2D;
+  *dist = R;
   
   } /* me_findjov */
